requirements.txt
./inaSpeechSegmenter
ffmpeg-python==0.2.0